papers/samarajeewaArtificialIntelligenceFramework2024/LEAP-Baseline-Forecast-Service/api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-servicebus
pyodbc
pandas
jsonify
numpy
scipy
scikit-learn
python-dotenv
xgboost==1.0.2
azure-storage-blob
joblib
dynaconf